auction: fuse both views in `dutch:actions::view`
